rename video2 -> video
