+++ rust/src/editor/mod.rs
@@ -1,4 +1,4 @@
-use crate::video2::Video;
+use crate::video::Video;
 use gpui::{
     App, Context, CursorStyle, DragMoveEvent, Entity, FocusHandle, KeyBinding, MouseButton,
     MouseDownEvent, MouseMoveEvent, MouseUpEvent, ObjectFit, PathPromptOptions, Render,
