+++ rust/src/editor/timeline_video.rs
@@ -2,7 +2,7 @@ use super::{
     clip_render_plan::resolve_visual_clip_render_plan, export::ExportOptions,
     export_gstreamer::build_timeline, timeline::Timeline,
 };
-use crate::video2::Video;
+use crate::video::Video;
 use ges::prelude::*;
 use gstreamer as gst;
 use gstreamer_app as gst_app;
